change aggregate durations to use uint32 throughout
This is related to changes in #925 and makes things more uniform while
we're using uint32 to represent second-granularity timestamps. If/when
we decide to move back to uint64, this change can be undone as well.

Note that changes to the proto types are also safe because the values
are variable-length encoded and therefore binary compatible:
https://developers.google.com/protocol-buffers/docs/proto#updating

+++ go/fennel/controller/aggregate/defs_offline_test.go
@@ -25,7 +25,7 @@ func TestDeactivateOffline(t *testing.T) {
 		Timestamp: 1,
 		Options: aggregate.Options{
 			AggType:      "topk",
-			Durations:    []uint64{3600 * 24 * 7},
+			Durations:    []uint32{3600 * 24 * 7},
 			CronSchedule: "20 1 * * ?",
 			Limit:        10,
 		},
